refactor(tests): test_merge_gate_phase_review.py をクラス単位で 3 ファイルに分割
644行を超過していたテストファイルを責務ごとに分割:
- test_phase_review_checkpoint.py: CheckpointPresence + SkipLabels (10 tests)
- test_phase_review_guard.py: CriticalFindings + ForceWarning (12 tests)
- test_merge_gate_integration.py: ExecuteIntegration (3 tests)
共通 fixture・ヘルパーを conftest.py に移動。全 25 テスト PASS 確認済み。

Closes #461
