+++ deltaspec/changes/issue-461/.deltaspec.yaml
@@ -1,0 +1,5 @@
+schema: spec-driven
+created: 2026-04-12
+issue: 461
+name: issue-461
+status: pending
